ml_pipeline/requirements.txt
# Core Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0  # YOLOv8 for instance segmentation

# Alternative: Mask R-CNN (if needed)
# mmdet>=3.0.0
# mmcv-full>=1.7.0

# Image Processing
opencv-python>=4.8.0
scikit-image>=0.21.0
scipy>=1.10.0
Pillow>=10.0.0
tifffile>=2023.0.0
albumentations>=1.3.0  # Advanced data augmentation

# Data Science & Visualization
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Experiment Tracking
mlflow>=2.8.0
# wandb>=0.15.0  # Alternative to MLflow

# Object Tracking
filterpy>=1.4.5  # Kalman filter for tracking
scikit-learn>=1.3.0  # For clustering and metrics
lap>=0.4.0  # Linear assignment problem solver

# Metrics & Evaluation
pycocotools>=2.0.6  # COCO evaluation metrics

# Annotation Handling
xmltodict>=0.13.0  # XML parsing
pyyaml>=6.0

# Utilities
tqdm>=4.65.0  # Progress bars
jupyterlab>=4.0.0  # For notebooks
ipywidgets>=8.0.0

# Optional: Super-resolution (if using Bayesian_DPA_TISR)
# timm>=0.9.0
# h5py>=3.7.0

# Development Tools
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
